requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==6.0
django-debug-toolbar==6.1.0
Faker==39.0.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
ruff==0.14.9
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.1
whitenoise==6.11.0
